deleting item updated to be done with permissions

+++ backend/src/resolvers/Mutation.js
@@ -47,9 +47,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. first find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
     // 2. check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permissions.some((permission) =>
+      ['ADMIN', 'DELETEITEM'].includes(permission)
+    );
+    if (!ownsItem && !hasPermission) {
+      throw new Error('You do not have permission to do that');
+    }
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
